chore: Remove the old My Collection artwork screen
