+++ .secrets.baseline
@@ -471,7 +471,7 @@
         "filename": "src/app/Scenes/MyCollection/Screens/ArtworkForm/MyCollectionArtworkForm.tests.tsx",
         "hashed_secret": "b5b44d59e3036fde34acae6a4ac3a669e27496ab",
         "is_verified": false,
-        "line_number": 488
+        "line_number": 487
       }
     ],
     "src/app/Scenes/MyCollection/utils/randomMyCollectionArtwork.ts": [
@@ -532,5 +532,5 @@
       }
     ]
   },
-  "generated_at": "2022-04-07T16:26:41Z"
+  "generated_at": "2022-04-21T12:58:25Z"
 }
