+++ src/app/Scenes/MyCollection/mutations/myCollectionUpdateArtwork.ts
@@ -18,7 +18,7 @@ export function myCollectionUpdateArtwork(
               ... on MyCollectionArtworkMutationSuccess {
                 artwork {
                   ...MyCollectionArtwork_sharedProps @relay(mask: false)
-                  ...OldMyCollectionArtwork_sharedProps @relay(mask: false)
+                  ...MyCollectionArtwork_sharedProps @relay(mask: false)
                 }
               }
               ... on MyCollectionArtworkMutationFailure {
