+++ src/app/navigation/routes.tsx
@@ -187,7 +187,6 @@ function getDomainMap(): Record<string, RouteMatcher[] | null> {
 
     addRoute("/my-collection", "MyCollection"),
     addRoute("/my-collection/artwork/:artworkSlug", "MyCollectionArtwork"),
-    addRoute("/my-collection/artwork-details/:artworkSlug", "MyCollectionArtworkFullDetails"),
     addRoute("/my-collection/artworks/new", "AddOrEditMyCollectionArtwork"),
     addRoute("/my-collection/artworks/:artworkID/edit", "AddOrEditMyCollectionArtwork"),
     addRoute(
